new/backup/back/requirements.txt
mysql.connector
mysql-connector-python
flask
flask_cors
